Remove accidentally sexual language
Gendering the software is a bit odd, and making it female is going to lead to awkward implications. One pair of sentences might lead a reader to believe that a joke is being made about "using" women (for sex), while the gymnast language might lead a reader to believe that the joke is being continued. These proposed changes make such an inference implausible, I think.

I recommend eliminating the gendered language entirely, which allows you to avoid anybody inferring anything that you didn't mean to imply. No need to get awkward turns of phrase get in the way of good software! :)